refactor: extract PortMixin to eliminate duplicated port properties
Fixes #22
